requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
frozenlist==1.3.3
h11==0.14.0
hiredis==2.2.2
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
nodeenv==1.7.0
prisma==0.8.1
pydantic==1.10.5
python-dotenv==0.21.1
rfc3986==1.5.0
sniffio==1.3.0
tomlkit==0.11.6
typing_extensions==4.5.0
yarl==1.8.2
